refactor(dhcp): extract DHCPv6 client to dhcpv6.go; finish the split
Complete the #6430 split by moving the DHCPv6 client out of dhcp.go
into a new sibling dhcpv6.go: the errV6AddrInvalidated sentinel, the
runDHCPv6() solicit / T1-renew / T2-rebind run loop, the dhcpv6Result
type, doDHCPv6() single-exchange primitive, selectIANAAddress() /
parseV6Reply() IA_NA/IA_PD reply parsing, buildDHCPv6Modifiers() /
buildDHCPv6RenewModifiers() DUID+ORO construction,
extractDelegatedPrefixes(), discoverIPv6Router(), waitForLinkLocal(),
and the DeriveSubPrefix() prefix-delegation helper.

Pure code-motion: whole functions, the dhcpv6Result type, and the
errV6AddrInvalidated var move verbatim with unchanged signatures and
bodies; only the per-file package clause and (goimports-pruned) import
block differ. dhcp.go is now the manager core alone — the value types,
Manager lifecycle/registry (Start / finishClient / Renew / StopAll /
Close), the option setters, and the shared netlink address plumbing
(applyAddress / removeAddress / scheduleRecompile / prefixToIPNet) —
shrinking from 2,148 to 529 lines.

Behavior-preservation proof: a normalized multiset diff of the original
dhcp.go against the concatenation of dhcp.go + lease.go + duid.go +
dhcpv4.go + dhcpv6.go reports 1373 == 1373 logic lines with zero delta.

Validation: go build ./pkg/dhcp/, go vet ./pkg/dhcp/, go test
./pkg/dhcp/ all clean.
